fix: fsd 위반 부분 리팩토링
