+++ src/features/laws/model/lawBottomSheetSchema.ts
@@ -1,0 +1,7 @@
+import { LawBottomSheetProps } from '@/features/laws/ui/LawBottomSheet';
+
+declare module '@/shared/store/bottomSheetStore' {
+  interface BottomSheetMap {
+    law: Omit<LawBottomSheetProps, 'isOpen' | 'onClose'>;
+  }
+}
